// packages/@coorpacademy-components/src/types/styles.ts
export type FlexDirection = 'column' | 'row' | 'row-reverse' | 'column-reverse' | undefined;
export type JustifyContent =
  | 'center'
  | 'flex-start'
  | 'flex-end'
  | 'space-between'
  | 'space-around'
  | 'space-evenly'
  | undefined;
export type Position = 'absolute' | 'relative' | undefined;
export type FlexAlignType = 'flex-start' | 'flex-end' | 'center' | 'stretch' | 'baseline';
export type AlignSelf = 'auto' | FlexAlignType | undefined;
export type Overflow = 'visible' | 'hidden' | 'scroll' | undefined;